Create fastParameters for icons and status table
