+++ packages/nimble-components/src/button/tests/button-matrix.stories.ts
@@ -4,13 +4,11 @@ import { pascalCase } from '@microsoft/fast-web-utilities';
 import { ButtonAppearance, ButtonAppearanceVariant } from '../types';
 import {
     createMatrix,
-    sharedMatrixParameters
+    sharedMatrixParameters,
+    createMatrixThemeStory
 } from '../../utilities/tests/matrix';
 import { disabledStates, DisabledState } from '../../utilities/tests/states';
-import {
-    createMatrixThemeStory,
-    createStory
-} from '../../utilities/tests/storybook';
+import { createStory } from '../../utilities/tests/storybook';
 import { hiddenWrapper } from '../../utilities/tests/hidden';
 import { textCustomizationWrapper } from '../../utilities/tests/text-customization';
 import { buttonTag } from '..';
