+++ packages/nimble-components/src/listbox/tests/listbox-matrix.stories.ts
@@ -2,12 +2,10 @@ import { ViewTemplate, html } from '@microsoft/fast-element';
 import type { Meta, StoryFn } from '@storybook/html';
 import {
     createMatrix,
-    sharedMatrixParameters
+    sharedMatrixParameters,
+    createMatrixThemeStory
 } from '../../utilities/tests/matrix';
-import {
-    createMatrixThemeStory,
-    createStory
-} from '../../utilities/tests/storybook';
+import { createStory } from '../../utilities/tests/storybook';
 import { listboxTag } from '..';
 import { listOptionTag } from '../../list-option';
 import { hiddenWrapper } from '../../utilities/tests/hidden';
