+++ packages/nimble-components/src/rich-text/viewer/tests/rich-text-viewer-matrix.stories.ts
@@ -1,10 +1,8 @@
 import type { Meta, StoryFn } from '@storybook/html';
 import { html, ViewTemplate } from '@microsoft/fast-element';
+import { createStory } from '../../../utilities/tests/storybook';
 import {
     createMatrixThemeStory,
-    createStory
-} from '../../../utilities/tests/storybook';
-import {
     createMatrix,
     sharedMatrixParameters
 } from '../../../utilities/tests/matrix';
