+++ packages/nimble-components/src/table-column/icon/tests/table-column-icon-matrix.stories.ts
@@ -1,7 +1,9 @@
 import type { StoryFn, Meta } from '@storybook/html';
 import { html, ViewTemplate } from '@microsoft/fast-element';
-import { createMatrixThemeStory } from '../../../utilities/tests/storybook';
-import { sharedMatrixParameters } from '../../../utilities/tests/matrix';
+import {
+    createMatrixThemeStory,
+    sharedMatrixParameters
+} from '../../../utilities/tests/matrix';
 import { Table, tableTag } from '../../../table';
 import { tableColumnIconTag } from '..';
 import { iconCheckTag } from '../../../icons/check';
