+++ packages/nimble-components/src/text-area/tests/text-area-matrix.stories.ts
@@ -1,15 +1,13 @@
 import type { StoryFn, Meta } from '@storybook/html';
 import { html, ViewTemplate } from '@microsoft/fast-element';
 import { pascalCase } from '@microsoft/fast-web-utilities';
+import { createStory } from '../../utilities/tests/storybook';
 import {
     createMatrixThemeStory,
-    createStory
-} from '../../utilities/tests/storybook';
-import { TextAreaAppearance } from '../types';
-import {
     createMatrix,
     sharedMatrixParameters
 } from '../../utilities/tests/matrix';
+import { TextAreaAppearance } from '../types';
 import {
     disabledStates,
     DisabledState,
